requirements.txt
theHarvester
python-whois
dnspython
python-nmap
requests
flask
